kazantsev/hw2/requirements.txt
numpy
transformers
scipy.special
streamlit